requirements.txt
streamlit
langchain
langchain-community
langchain-openai
mysql-connector-python
python-dotenv
